* origin/released:
  xserver libxcb libxcb/xcb-proto fontconfig mesa putty git update Jul 24 2017

+++ mesalib/src/gallium/auxiliary/hud/hud_fps.c
@@ -47,12 +47,12 @@ query_fps(struct hud_graph *gr)
 
    if (info->last_time) {
       if (info->last_time + gr->pane->period <= now) {
-         double fps = (uint64_t)info->frames * 1000000 /
+         double fps = ((uint64_t)info->frames) * 1000000 /
                       (double)(now - info->last_time);
          info->frames = 0;
          info->last_time = now;
 
-         hud_graph_add_value(gr, (uint64_t) fps);
+         hud_graph_add_value(gr, fps);
       }
    }
    else {
